requirements.txt
astroid==2.1.0
autopep8==1.4.3
bcrypt==3.1.6
cffi==1.12.1
Django==1.11.17
django-cleanup==3.2.0
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.4.1
pycodestyle==2.5.0
pycparser==2.19
pylint==2.2.2
pytz==2018.9
six==1.12.0
urllib3==1.24.1
wrapt==1.11.1
